# backend/src/main/kotlin/com/base/wealth/domain/model/Holding.kt
package com.base.wealth.domain.model

data class Holding(
    val id: Long,
    val name: String,
    val cls: AssetClass,
    val platform: String,
    val qty: Double? = null,
    val value: Double, // Almacenado en moneda base USD
    val currency: Currency,
    val change: Double // Variación 24h en %
)